HHH-9261 :  Backport HHH-9106 to 4.2 (Multiple representations of the same entity cannot be merged using cascade=merge)
