+++ hibernate-entitymanager/src/test/java/org/hibernate/ejb/test/emops/MergeMultipleEntityCopiesAllowedLoggedTest.java
@@ -21,7 +21,7 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.hibernate.jpa.test.emops;
+package org.hibernate.ejb.test.emops;
 
 import java.util.Map;
