+++ hibernate-entitymanager/src/test/java/org/hibernate/ejb/test/emops/MergeMultipleEntityCopiesAllowedTest.java
@@ -21,29 +21,38 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.hibernate.jpa.test.emops;
+package org.hibernate.ejb.test.emops;
 
 import java.util.List;
-
+import java.util.Map;
 import javax.persistence.EntityManager;
 
 import org.junit.Test;
 
-import org.hibernate.jpa.test.BaseEntityManagerFunctionalTestCase;
+import org.hibernate.ejb.test.BaseEntityManagerFunctionalTestCase;
 import org.hibernate.testing.TestForIssue;
 
-import static junit.framework.TestCase.fail;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertSame;
 
 /**
- * Tests merging multiple detached representations of the same entity when
- * not allowed by default.
+ * Tests merging multiple detached representations of the same entity when it is explicitly allowed.
  *
  * @author Gail Badner
  */
 @TestForIssue( jiraKey = "HHH-9106")
-public class MergeMultipleEntityCopiesDisallowedByDefaultTest extends BaseEntityManagerFunctionalTestCase {
+public class MergeMultipleEntityCopiesAllowedTest extends BaseEntityManagerFunctionalTestCase {
+
+	@SuppressWarnings( {"unchecked"})
+	protected void addConfigOptions(Map options) {
+		super.addConfigOptions( options );
+		options.put(
+				"hibernate.event.merge.entity_copy_observer",
+				"allow"
+		);
+	}
 
 	@Test
 	public void testCascadeFromDetachedToNonDirtyRepresentations() {
@@ -60,7 +69,7 @@ public class MergeMultipleEntityCopiesDisallowedByDefaultTest extends BaseEntity
 		em.getTransaction().commit();
 		em.close();
 
-		// Get another representation of the same Item from a different session.
+		// Get another representation of the same Item from a different EntityManager.
 
 		em = getOrCreateEntityManager();
 		Item item1_1 = em.find( Item.class, item1.getId() );
@@ -76,20 +85,27 @@ public class MergeMultipleEntityCopiesDisallowedByDefaultTest extends BaseEntity
 
 		em = getOrCreateEntityManager();
 		em.getTransaction().begin();
-		try {
-			em.merge( hoarder );
-			fail( "should have failed due IllegalStateException");
-		}
-		catch (IllegalStateException ex) {
-			//expected
-		}
-		finally {
-			em.getTransaction().rollback();
-			em.close();
-		}
+		hoarder = em.merge( hoarder );
+		assertEquals( 1, hoarder.getItems().size() );
+		assertSame( hoarder.getFavoriteItem(), hoarder.getItems().iterator().next() );
+		assertEquals( item1.getId(), hoarder.getFavoriteItem().getId() );
+		assertEquals( item1.getCategory(), hoarder.getFavoriteItem().getCategory() );
+		em.getTransaction().commit();
+		em.close();
+
+		em = getOrCreateEntityManager();
+		em.getTransaction().begin();
+		hoarder = em.merge( hoarder );
+		assertEquals( 1, hoarder.getItems().size() );
+		assertSame( hoarder.getFavoriteItem(), hoarder.getItems().iterator().next() );
+		assertEquals( item1.getId(), hoarder.getFavoriteItem().getId() );
+		assertEquals( item1.getCategory(), hoarder.getFavoriteItem().getCategory() );
+		em.getTransaction().commit();
+		em.close();
 
 		cleanup();
 	}
+
 
 	@Test
 	public void testTopLevelManyToOneManagedNestedIsDetached() {
@@ -121,17 +137,19 @@ public class MergeMultipleEntityCopiesDisallowedByDefaultTest extends BaseEntity
 		category.setExampleItem( item1_1 );
 
 		// now item1Merged is managed and it has a nested detached item
-		try {
-			em.merge( item1Merged );
-			fail( "should have failed due IllegalStateException");
-		}
-		catch (IllegalStateException ex) {
-			//expected
-		}
-		finally {
-			em.getTransaction().rollback();
-			em.close();
-		}
+		em.merge( item1Merged );
+		assertEquals( category.getName(), item1Merged.getCategory().getName() );
+		assertSame( item1Merged, item1Merged.getCategory().getExampleItem() );
+		em.getTransaction().commit();
+		em.close();
+
+		em = getOrCreateEntityManager();
+		em.getTransaction().begin();
+		item1 = em.find( Item.class, item1.getId() );
+		assertEquals( category.getName(), item1.getCategory().getName() );
+		assertSame( item1, item1.getCategory().getExampleItem() );
+		em.getTransaction().commit();
+		em.close();
 
 		cleanup();
 	}
