--- hibernate-entitymanager/src/test/java/org/hibernate/jpa/test/emops/MergeMultipleEntityCopiesDisallowedTest.java
@@ -1,47 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * Copyright (c) 2014, Red Hat Inc. or third-party contributors as
- * indicated by the @author tags or express copyright attribution
- * statements applied by the authors.  All third-party contributions are
- * distributed under license by Red Hat Inc.
- *
- * This copyrighted material is made available to anyone wishing to use, modify,
- * copy, or redistribute it subject to the terms and conditions of the GNU
- * Lesser General Public License, as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
- * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
- * for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this distribution; if not, write to:
- * Free Software Foundation, Inc.
- * 51 Franklin Street, Fifth Floor
- * Boston, MA  02110-1301  USA
- */
-package org.hibernate.jpa.test.emops;
-
-import java.util.Map;
-
-import org.hibernate.testing.TestForIssue;
-
-/**
- * Tests merging multiple detached representations of the same entity when
- * explicitly disallowed.
- *
- * @author Gail Badner
- */
-@TestForIssue( jiraKey = "HHH-9106")
-public class MergeMultipleEntityCopiesDisallowedTest extends MergeMultipleEntityCopiesDisallowedByDefaultTest {
-
-	@SuppressWarnings( {"unchecked"})
-	protected void addConfigOptions(Map options) {
-		super.addConfigOptions( options );
-		options.put(
-				"hibernate.event.merge.entity_copy_observer",
-				"disallow"
-		);
-	}
-}
